Refactored code. socketRoomManager now handles the game logic.
